C64 updates
* Update to VICE 2.4.28
* Add Virtual Device Traps option
* Add more options to System & Media Control menu including those for disk drives
* Set AutostartPrgDisk.d64 as PRG autostart filename
* Set TDE default to off
* Various menu item tweaks

+++ C64.emu/src/vice/platform/platform_linux_runtime_os.c
@@ -109,6 +109,7 @@
    x86          | glibc-2.17
    x86          | glibc-2.18
    x86          | glibc-2.19
+   x86          | glibc-2.21
    x86          | dietlibc
    x86          | newlib
    x86          | musl
@@ -118,6 +119,8 @@
    x86          | unixware (LKP)
    x86          | solaris (lxrun)
    x86          | netbsd (emulation layer)
+   x86          | freebsd (emulation layer)
+   x86          | openbsd (emulation layer)
  */
 
 #include "vice.h"
@@ -130,6 +133,7 @@
 #include <sys/utsname.h>
 #include <string.h>
 #include <unistd.h>
+#include <sys/stat.h>
 
 #include "archdep.h"
 #include "lib.h"
@@ -168,8 +172,8 @@ char *platform_get_linux_runtime_cpu(void)
             tempfile = archdep_tmpnam();
             tempsystem = util_concat("cat /proc/cpuinfo >", tempfile, NULL);
             if (system(tempsystem) < 0) {
-		log_warning(LOG_ERR, "`%s' failed.", tempsystem);
-	    }
+                log_warning(LOG_ERR, "`%s' failed.", tempsystem);
+            }
             cpuinfo = fopen(tempfile, "rb");
         }
         if (cpuinfo) {
@@ -248,24 +252,29 @@ char *platform_get_linux_runtime_os(void)
     struct utsname name;
     FILE *bsd_emul_test;
     int is_bsd = 0;
+    int ret;
+    int i = 0;
 
     if (!got_linux_version) {
         unlink("emultest.sh");
+        unlink("emultest.netbsd");
         bsd_emul_test = fopen("emultest.sh", "wb");
         if (bsd_emul_test) {
-            unlink("emultest.result");
             fprintf(bsd_emul_test, "#!/bin/sh\n");
             fprintf(bsd_emul_test, "if test -f /proc/self/emul; then\n");
-            fprintf(bsd_emul_test, "  echo emulation >emultest.result\n");
+            fprintf(bsd_emul_test, "  echo emulation >emultest.netbsd\n");
             fprintf(bsd_emul_test, "fi\n");
             fclose(bsd_emul_test);
-            system("sh ./emultest.sh");
-            unlink("emultest.sh");
-            bsd_emul_test = fopen("emultest.result", "rb");
+            chmod("emultest.sh", S_IRWXU);
+            ret = system("./emultest.sh");
+            if (!ret) {
+                unlink("emultest.sh");
+            }
+            bsd_emul_test = fopen("emultest.netbsd", "rb");
             if (bsd_emul_test) {
                 sprintf(linux_version, "NetBSD");
                 fclose(bsd_emul_test);
-                unlink("emultest.result");
+                unlink("emultest.netbsd");
                 is_bsd = 1;
             }
         }
@@ -281,6 +290,12 @@ char *platform_get_linux_runtime_os(void)
                 } else {
                     sprintf(linux_version, "lxrun sco");
                 }
+            } else if (!strncasecmp(name.version, "FreeBSD", 7)) {
+                while (name.version[i] != '-' && name.version[i] != 0) {
+                    linux_version[i] = name.version[i];
+                    i++;
+                }
+                linux_version[i] = 0;
             } else {
                 sprintf(linux_version, "%s %s", name.sysname, name.release);
             }
