+++ C64.emu/src/vice/platform/platform_solaris_runtime_os.c
@@ -25,25 +25,40 @@
  */
 
 /* Tested and confirmed working on:
- *
- * Solaris 2.3 (sparc)
- * Solaris 2.4 (intel)
- * Solaris 2.4 (sparc)
- * Solaris 2.5.1 (intel)
- * Solaris 2.5.1 (sparc)
- * Solaris 2.6 (intel)
- * Solaris 2.6 (sparc)
- * Solaris 7 (intel)
- * Solaris 7 (sparc)
- * Solaris 8 (intel)
- * Solaris 8 (sparc)
- * Solaris 9 (intel)
- * Solaris 9 (sparc)
- * Solaris 10 (intel 32&64)
- * Solaris 10 (sparc)
- * OpenSolaris (intel)
- * Solaris 11(.0) (intel 32&64)
- * Solaris 11.1 (intel 32&64)
+   cpu     | Operating System
+   --------------------------
+   sparc   | Solaris 2.2
+   sparc   | Solaris 2.3
+   intel   | Solaris 2.4
+   sparc   | Solaris 2.4
+   intel   | Solaris 2.5.1
+   ppc     | Solaris 2.5.1
+   sparc   | Solaris 2.5.1
+   intel   | Solaris 2.6
+   sparc   | Solaris 2.6
+   intel   | Solaris 7
+   sparc   | Solaris 7
+   sparc64 | Solaris 7
+   intel   | Solaris 8
+   sparc   | Solaris 8
+   sparc64 | Solaris 8
+   intel   | Solaris 9
+   sparc   | Solaris 9
+   sparc64 | Solaris 9
+   intel   | Solaris 10
+   sparc   | Solaris 10
+   sparc64 | Solaris 10
+   x86     | OpenSolaris
+   x64     | OpenSolaris
+   x86     | Solaris 11
+   x64     | Solaris 11
+   x86     | Solaris 11.1
+   x64     | Solaris 11.1
+   x86     | Solaris 11.2
+   x64     | Solaris 11.2
+   x86     | Solaris 11.3
+   x64     | Solarid 11.3
+   i386    | NetBSD (emulation layer)
  */
 
 #include "vice.h"
@@ -84,17 +99,25 @@ char *platform_get_solaris_runtime_os(void)
                 os = "Solaris 11";
             } else if (!strcasecmp(name.version, "11.1")) {
                 os = "Solaris 11.1";
-            else {
+            } else if (!strcasecmp(name.version, "11.2")) {
+                os = "Solaris 11.2";
+            } else if (!strcasecmp(name.version, "11.3")) {
+                os = "Solaris 11.3";
+            } else {
                 os = "OpenSolaris";
             }
         } else {
-            os = "Unknown Solaris version";
+            if (!strcasecmp(name.sysname, "NetBSD")) {
+                os = "NetBSD";
+            } else {
+                os = "Unknown Solaris version";
+            }
         }
     }
     return os;
 }
 
-#if defined(__sparc64__) || defined(sparc64) || defined(__sparc__) || defined(sparc)
+#if defined(__sparc64__) || defined(sparc64) || defined(__sparc__) || defined(sparc) || defined(__PPC__) || defined(__ppc)
 #include <sys/types.h>
 #include <sys/processor.h>
 #include <stdio.h>
@@ -105,7 +128,7 @@ static int got_cpu = 0;
 char *platform_get_solaris_runtime_cpu(void)
 {
     processor_info_t info;
-    int status = 0;
+    int status = -1;
     struct utsname name;
     FILE *infile = NULL;
     size_t size = 0;
@@ -115,36 +138,40 @@ char *platform_get_solaris_runtime_cpu(void)
     char *loc2 = NULL;
 
     if (!got_cpu) {
-        status = processor_info(0, &info);
+        uname(&name);
+        if (strcasecmp(name.sysname, "NetBSD")) {
+            status = processor_info(0, &info);
+        }
         if (status != -1) {
             sprintf(solaris_cpu, "%s", info.pi_processor_type);
         } else {
-            uname(&name);
             sprintf(solaris_cpu, "%s", name.machine);
         }
-        system("dmesg >/tmp/vice.cpu.tmp");
-        infile = fopen("/tmp/vice.cpu.tmp", "rb");
-        if (infile) {
-            fseek(infile, 0L, SEEK_END);
-            size = ftell(infile);
-            fseek(infile, 0L, SEEK_SET);
-            buffer = (char *)malloc(size);
-            size2 = fread(buffer, 1, size, infile);
-            if (size == size2) {
-                loc = strstr(buffer, "cpu0:");
-                if (loc) {
-                    loc += 6;
-                    loc2 = strstr(loc, " (");
-                    if (loc2) {
-                        *loc2 = 0;
-                        sprintf(solaris_cpu, "%s (%s)", solaris_cpu, loc);
+        if (strcasecmp(name.sysname, "NetBSD")) {
+            system("dmesg >/tmp/vice.cpu.tmp");
+            infile = fopen("/tmp/vice.cpu.tmp", "rb");
+            if (infile) {
+                fseek(infile, 0L, SEEK_END);
+                size = ftell(infile);
+                fseek(infile, 0L, SEEK_SET);
+                buffer = (char *)malloc(size);
+                size2 = fread(buffer, 1, size, infile);
+                if (size == size2) {
+                    loc = strstr(buffer, "cpu0:");
+                    if (loc) {
+                        loc += 6;
+                        loc2 = strstr(loc, " (");
+                        if (loc2) {
+                            *loc2 = 0;
+                            sprintf(solaris_cpu, "%s (%s)", solaris_cpu, loc);
+                        }
                     }
                 }
+                fclose(infile);
+                free(buffer);
             }
-            fclose(infile);
-            free(buffer);
+            unlink("/tmp/vice.cpu.tmp");
         }
-        unlink("/tmp/vice.cpu.tmp");
         got_cpu = 1;
     }
     return solaris_cpu;
